complete EA framework.

+++ Evaluator/Utils/dataset.py
@@ -53,7 +53,7 @@ def build_cifar10(data_path,
                 num_worker=10,
                 train_batch_size=32,
                 eval_batch_size=32,
-                split_train_for_valid:int=None, **kwargs):
+                split_train_for_valid:float=None, **kwargs):
 
     train_transform, valid_transform = _data_transforms_cifar10(cutout_size)
     if split_train_for_valid is None:
